Tests/LibWeb/Text/expected/css/style-invalidation/input-type-same-state-no-invalidation.txt

FAIL: same-state type set recomputed 504 elements
